Permitir cambio de saldo mayor al de la unidad proveedores

+++ app/website/controllers/apiProveedor.js
@@ -99,7 +99,7 @@ Apiproveedor.prototype.get_ProveedorPolizaDetalle = function(req, res, next) {
     { name: 'idfinancieraD', value: req.query.idfinancieraD, type: self.model.types.INT },
     { name: 'idEsquemaD', value: req.query.idEsquemaD, type: self.model.types.INT },
     { name: 'idUsuario', value: req.query.idUsuario, type: self.model.types.INT },
-    { name: 'montoFinanciar', value: req.query.montoFinanciar, type: self.model.types.INT },
+    { name: 'montoFinanciar', value: req.query.montoFinanciar, type: self.model.types.DECIMAL },
     { name: 'fechaInicio', value: req.query.fechaInicio, type: self.model.types.STRING }];
 
     self.model.query('Pol_Poliza4Detalle_INS', params, function(error, result) {
